Implement Day 1 in Kotlin
For Part 1, I accidentally was doing modulo of 99 instead of 100 since I misread the problem, but it was an easy enough fix.

I tried to use math on the quotient to calculate the answer to Part 2 in a "smart" way, but I kept getting wrong answers so I switched to a simpler solution.