Implement Day 6 in Kotlin
I was able to get Part 1 pretty quickly, in 3 minutes and 43 seconds to be exact. However, Part 2 gave me a bit more trouble. Originally, I was splitting based on the regex " +" for Part 1, meaning I was now losing important alignment information. I first tried to hack the regex to fit the new purpose with "(?!^) +(?!$)" since I was getting index-out-of-bound exceptions at the first column but then realized my mistake. I switched to a more manual parsing approach, and then I had to reason about how to construct the vertical numbers properly. In total, I took 14 minutes and 48 seconds after solving Part 1 to solve Part 2.

One thing that was amusing to me about the problem statement for Part 2 is the stipulation of right-to-left column evaluation order, but for commutative operations like addition and multiplication (the only supported operations), this stipulation is meaningless. I ignored implementing it during my live solution and in the final committed code since it doesn't affect the answer whatsoever.