Implement Day 7 in Kotlin
I had some other things I was trying to do at puzzle open, but I took a look at Part 1, and it seemed simple enough, so I did it quickly in about 5 minutes after puzzle unlock. I then took about a 15-minute break to finish up some things and returned to do Part 2. I had the right logic for Part 2 after about another 5 minutes, but a pesky bug meant I was losing a few timelines if a splitter had a split beam that coincided with an existing beam; in that case, one of the two would be overwritten and lost. This meant I was losing 5 timelines on the example input and getting an answer of 35 instead of 40. I spent a good bit of time (about 20 minutes) trying to debug before I realized what had gone wrong. It ended up being a very quicky and easy fix.

I then translated all of my logic into a one-liner just for fun. Even with all of the slowness of Kotlin and especially with convoluted one-liners, it runs in about 80 milliseconds on my machine!

Also, because of the grid-like design of the puzzle and the relative simplicity of the solution, I decided to challenge myself to implementing a solver for Day 7 in Google Sheets. After about an hour or two of work, I got a serviceable solution! Here is the template if you would like to try it out (simply paste your input in the "Input" sheet, then wait for the answer to appear in the "Output" sheet): https://docs.google.com/spreadsheets/d/18PBwKAv-llardcSQA5SUmB-5BuNjKGhMd83hWmJpFCM/edit?usp=sharing