Implement Day 9 in Kotlin
Part 1 was pretty trivial; I finished it in about 3 minutes. Part 2 was a lot trickier. I thought that a point-in-polygon style approach would work pretty early on, but went about implementing it in a very naive way. After a few hours on and off of fiddling with it, I was able to come up with the idea of grouping ranges by y-value and for each rectangle, casting horizontal rays through each y-value and seeing if it is inside the shape. In the end, I did use a point-in-polygon approach (sort of)!

My code is not fast by any means; it takes about 10 seconds to run on my hardware. But I don't feel like thinking of a better solution right now.