Optimize Day 9 in Kotlin
This idea isn't mine, but it's silly that I didn't think of it myself. Similar to how I "compress" the horizontal ranges so that I don't actually have to iterate every single x-value, one can compress the vertical range to only check y-values that line up with a red tile, since between the y-values of 2 consecutive red tiles (sorted vertically), the shape cannot change.

The optimization is trivial to implement and drops the runtime from 11 seconds to 250 milliseconds on my hardware.