Implement Day 10 in Kotlin
For Part 1, I tried to overoptimize and think of a complicated solution before giving up and landing on a simple DFS that runs in 114 milliseconds. For Part 2, I couldn't think of a nice way to solve it, so I used Z3. It took a bit of work, because I didn't remember the specifics of how to use Z3, but I was also trying to figure it out myself and look up minimal resources. Looking back on it, this is a trivial linear algebra problem, so I want to switch to making a LinearAlgebra utility class with my own math/work to help with this problem. Including Z3 for Part 2, the whole thing runs in about 4-5 seconds on my hardware.