Implement Day 12 in Kotlin
I didn't really like today's puzzle. The solution ended up being characterized by a simple precondition that only worked on the real input and not on the provided example; specifically, checking that the total number of required filled-in spots by summing over the presents would even possibly fit in the total number of spots under the given tree. From talking with others, it *appears* like the problem in complete generality is NP-complete. So it is understandable why it would be infeasible to check all the possible options for each tree. However, the fact that the solution ended up being a simple math check that basically removes all the challenge and intrigue of the puzzle frustrates me.

When the problem description for Day 12 opened, I was pretty stumped on how to approach this in a way that was tractable. I stared at it for a little while and resigned myself to trying it later. However, I was pretty shocked to see that there was already dozens of people solving it within the first few minutes it released. The puzzle, at least at first glance, reminded me of something akin to [Jurrasic Jigsaw](https://adventofcode.com/2020/day/20) from AOC 2020 day 20, which was quite a difficult puzzle. During the course of writing my solution, I thought of the precondition as a simple stepping stone to the final solution. Once I fixed the bugs in my (now disabled) search function, I noticed that trees which satisfied this precondition completed very quickly, and trees that didn't would basically never finish, using my crude search. This led me to getting the answer, but I was dissatisfied to find out that all the code for piece rotation and flipping plus searching for a possible tree configuration was entirely unnecessary.

In the end, the puzzle boils down to input parsing plus a basic inequality for each tree. I guess I should have expected that, as in prior years, the difficulty of the final day's puzzle tends to be lowered a bit to serve as a palette cleanser and nice finish to the year. However, I guess I presumed that the puzzle was uniquely difficult since it was the final day of the first year that had been cut in half by length.

Anyways, I've left my search code for posterity, but it doesn't actually contribute anything meaningful anymore except adding to the runtime of the solution. I perhaps wish that the problem/input could have been modified somewhat to either make a variant that is more feasible to solve with a polynomial-time algorithm or by making the input small enough that current best-known algorithms would be feasible to implement and run.

It's sad knowing that this year's puzzles have already come to a close. I didn't feel like there was a particularly hard puzzle that "stood out" to me this year. Perhaps that is because there was half as many, or perhaps that is because I've had a lot of practice in prior years. I somewhat tried to do the puzzles when they released, but without a global leaderboard, I felt less motivated to do them immediately. Overall, it was still fun. Time will tell if I make it back for next year!