Clean up Day 10 slightly
After discussing with another person, they pointed out how a quite obvious upper bound on the maximum presses for a button is given directly by the input. (The example they gave was if a button's wiring was (0,1) and the joltages for 0 and 1 were 10 and 15, respectively, then that button's maximum possible value would be 10.) Using this on the buttons corresponding to the free variables let me remove the arbitrary cutoff without much issue.

However, this change meant I was processing more states than necessary, increasing the runtime by about 5x on my hardware. I realized there are a few more cases when I can quit early from searching further down impossible paths; specifically when the coefficients for the remaining free buttons are all the same sign and can't possibly satisfy the given equation. This change sped up my code, so now it is approximately 2x slower than the hardcoded cutoff from before.

I think this slowdown is perfectly acceptable, as the cutoff I chose was entirely arbitrary and a guess on the structure of the input, which felt like the wrong move.