Add small Day 10 optimization
Also, overhauled Part 1 to use bitmasks. The idea wasn't mine, but I saw it in others' solutions while working on Part 2. The core idea is that pushing a button twice is equivalent to not pushing it at all for Part 1. I had a cruder version of this idea in my DFS where I noticed that pushing the same button twice in a row was useless, but I didn't consider this more general case. Using this idea and bitmasks speeds up the runtime a little bit.